requirements.txt
discord.py==2.3.2
aiohttp==3.8.6
aiosignal==1.3.1
frozenlist==1.4.0
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.3.1
frozenlist==1.4.0
multidict==6.0.4
yarl==1.9.2
idna==3.4
multidict==6.0.4
PyNaCl==1.5.0
cffi==1.16.0
pycparser==2.21